created 2 more html pages and more
I created project.html and contact.html and I added the proper title for each page. I also added the <link> element in the <head> section of each HTML file to reference the 'styles.css' stylesheet. This was done to ensure that my CSS styles are applied to all three HTML pages.